WIP: Agregando Iseed y seed de usuarios de prueba

+++ database/migrations/2019_10_06_055455_create_twitt_states_table.php
@@ -15,7 +15,7 @@ class CreateTwittStatesTable extends Migration
     {
         Schema::create('twitt_states', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('tid');
+            $table->bigInteger('tid');
             $table->timestamps();
         });
     }
